+++ tests/tests.rs
@@ -4,31 +4,46 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-#[cfg(feature = "std")]
 mod std_tests {
     use core::convert::TryFrom;
     use dusk_pki::{PublicSpendKey, SecretSpendKey, ViewKey};
+    use rand::SeedableRng;
+
+    fn ssk_from_str(s: &str) -> SecretSpendKey {
+        use rand::rngs::StdRng;
+        use sha2::{Digest, Sha256};
+
+        let bytes = s.as_bytes();
+
+        let mut hasher = Sha256::default();
+        hasher.input(bytes);
+        let bytes = hasher.result();
+
+        let mut seed = [0u8; 32];
+        seed.copy_from_slice(&bytes[..32]);
+
+        SecretSpendKey::random(&mut StdRng::from_seed(seed))
+    }
 
     #[test]
     fn ssk_from_bytes() {
-        let bytes = b"some bytes".to_vec();
-
-        let ssk_a = SecretSpendKey::from(&bytes[..]);
-        let ssk_b = SecretSpendKey::from(&bytes[..]);
+        let ssk_a = ssk_from_str("some bytes");
+        let ssk_b = ssk_from_str("some bytes");
 
         assert_eq!(ssk_a, ssk_b);
     }
 
     #[test]
     fn keys_encoding() {
-        let bytes = b"some bytes".to_vec();
-
-        let ssk = SecretSpendKey::from(bytes.as_slice());
+        let ssk = ssk_from_str("some bytes");
         let vk = ssk.view_key();
         let psk = ssk.public_key();
 
-        assert_eq!(vk, ViewKey::try_from(format!("{}", vk)).unwrap());
-        assert_eq!(psk, PublicSpendKey::try_from(format!("{}", psk)).unwrap());
+        assert_eq!(vk, ViewKey::try_from(format!("{}", vk).as_str()).unwrap());
+        assert_eq!(
+            psk,
+            PublicSpendKey::try_from(format!("{}", psk).as_str()).unwrap()
+        );
     }
 
     #[test]
